Move FileMap to common module
