+++ tool/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod c;
 pub mod c2;
+pub mod common;
 pub mod cpp;
+mod docs_util;
 pub mod dotnet;
 pub mod js;
-
-mod docs_util;
 mod layout;
 mod util;
